Version 2024.10.18 (#8)

+++ system.c
@@ -9,7 +9,7 @@
 int qKey() { return _kbhit(); }
 int key() { return _getch(); }
 void ttyMode(int isRaw) {}
-void ms(cell sleepForMS) { Sleep(sleepForMS); }
+void ms(cell sleepForMS) { Sleep((DWORD)sleepForMS); }
 
 #endif
